requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Image handling
Pillow>=10.0.0

# Password hashing
bcrypt==4.1.2

# API Documentation
drf-yasg==1.21.7

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Excel/CSV export
openpyxl==3.1.2
pandas==2.2.0

# Email
django-anymail==10.2

# File storage
boto3==1.34.34

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==7.0.0
black==24.1.1

